+++ ranker/ranker.py
@@ -6,11 +6,6 @@ sys.path.append('..')
 sys.path.append('../crawler')
 import helpers
 log = helpers.log
-import numpy as np 
-
-import nltk
-from nltk.tokenize import word_tokenize
-from nltk.stem import PorterStemmer 
 
 from mongoengine import *
 from mongoConfig import *
@@ -20,6 +15,8 @@ app = Flask(__name__)
 
 port = 8002
 connect('chefmateDB', host='18.222.251.5', port=27017)
+
+inMemoryTFIDF = loadInvertedIndexToMemory()
 
 @app.route('/', methods=["GET"])
 def index():
@@ -39,29 +36,6 @@ def testRoute():
       1, 'successfully got packet from ranker', {'name': 'Ashwin'})
 
 #Insert Helper functions below here
-def loadInvertedIndexToMemory():
-  log('info', 'Loading Inverted Index into main memory.')
-  invertedIndex = InvertedIndex.objects()
-  inMemoryTFIDF= np.zeros((InvertedIndex.objects.count(), Crawler.objects.count()))
-
-  for termEntry in invertedIndex:
-    termNum = termEntry['termNum']
-    docInfoList = termEntry['doc_info']
-    for doc in docInfoList:
-      docId = int(doc['docId'])
-      inMemoryTFIDF[termNum][docId]=termEntry['tfidf'][str(docId)]
-
-  return inMemoryTFIDF
-
-def stemQuery(query):
-  output=[]
-  porterStemmer = PorterStemmer()
-  tokenedText = word_tokenize(query)
-  for word in tokenedText:
-    output.append(porterStemmer.stem(word))
-  return output
-
-inMemoryTFIDF = loadInvertedIndexToMemory()
 
 if __name__ == "__main__":
   log('info', f"Ranker is listening on port {port}, {app.config['ENV']} environment.")
